Make MEM and Memory driver have both raster, vector and multidim capabilities
