+++ .github/workflows/ubuntu_24.04/expected_gdalinfo_formats.txt
@@ -22,7 +22,7 @@ Supported Formats: (ro:read-only, rw:read-write, +:write from scratch, u:update,
   DTED -raster- (rwv): DTED Elevation Raster (*.dt0, *.dt1, *.dt2)
   PNG -raster- (rwv): Portable Network Graphics (*.png)
   JPEG -raster- (rwv): JPEG JFIF (*.jpg, *.jpeg)
-  MEM -raster,multidimensional raster- (rw+): In Memory Raster
+  MEM -raster,multidimensional raster,vector- (rw+): In Memory raster, vector and multidimensional raster
   JDEM -raster- (rov): Japanese DEM (.mem) (*.mem)
   GIF -raster- (rwv): Graphics Interchange Format (.gif) (*.gif)
   BIGGIF -raster- (rov): Graphics Interchange Format (.gif) (*.gif)
@@ -131,6 +131,7 @@ Supported Formats: (ro:read-only, rw:read-write, +:write from scratch, u:update,
   STACIT -raster- (rovs): Spatio-Temporal Asset Catalog Items
   JPEGXL -raster- (rwv): JPEG-XL (*.jxl)
   GDALG -raster,vector- (rov): GDAL Streamed Algorithm driver (*.gdalg.json)
+  Memory -raster,multidimensional raster,vector- (rw+): In Memory raster, vector and multidimensional raster
   GPKG -raster,vector- (rw+uvs): GeoPackage (*.gpkg, *.gpkg.zip)
   SQLite -raster,vector- (rw+uv): SQLite / Spatialite / RasterLite2 (*.sqlite, *.db)
   OpenFileGDB -raster,vector- (rw+uv): ESRI FileGeodatabase (using OpenFileGDB) (*.gdb)
