+++ frmts/mem/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_gdal_driver(
   TARGET gdal_MEM
-  SOURCES memdataset.cpp memdataset.h
+  SOURCES memdataset.cpp memdataset.h ogrmemlayer.cpp
   BUILTIN)
 gdal_standard_includes(gdal_MEM)
