+++ frmts/stacta/CMakeLists.txt
@@ -1,3 +1,2 @@
 add_gdal_driver(TARGET gdal_STACTA SOURCES stactadataset.cpp STRONG_CXX_WFLAGS PLUGIN_CAPABLE NO_DEPS)
 gdal_standard_includes(gdal_STACTA)
-target_include_directories(gdal_STACTA PRIVATE ${GDAL_RASTER_FORMAT_SOURCE_DIR}/mem)
