+++ frmts/zarr/CMakeLists.txt
@@ -24,8 +24,7 @@ if(NOT TARGET gdal_Zarr)
 endif()
 
 gdal_standard_includes(gdal_Zarr)
-target_include_directories(gdal_Zarr PRIVATE ${GDAL_RASTER_FORMAT_SOURCE_DIR}/mem
-                                             ${GDAL_RASTER_FORMAT_SOURCE_DIR}/netcdf)
+target_include_directories(gdal_Zarr PRIVATE ${GDAL_RASTER_FORMAT_SOURCE_DIR}/netcdf)
 
 if (GDAL_USE_BLOSC)
   target_compile_definitions(gdal_Zarr PRIVATE -DHAVE_BLOSC)
