+++ ogr/ogr_geocoding.cpp
@@ -26,7 +26,7 @@
 #include "ogr_core.h"
 #include "ogr_feature.h"
 #include "ogr_geometry.h"
-#include "ogr_mem.h"
+#include "memdataset.h"
 #include "ogrsf_frmts.h"
 
 // Emulation of gettimeofday() for Windows.
