+++ ogr/ogrsf_frmts/geojson/ogr_geojson.h
@@ -16,7 +16,7 @@
 
 #include "cpl_port.h"
 #include "ogrsf_frmts.h"
-#include "../mem/ogr_mem.h"
+#include "memdataset.h"
 
 #include <cstdio>
 #include <vector>  // Used by OGRGeoJSONLayer.
