+++ ogr/ogrsf_frmts/mem/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_gdal_driver(
   TARGET ogr_MEM
-  SOURCES ogrmemdatasource.cpp ogr_mem.h ogrmemdriver.cpp ogrmemlayer.cpp
+  SOURCES ogrmemdriver.cpp
   BUILTIN)
 gdal_standard_includes(ogr_MEM)
