+++ ogr/ogrsf_frmts/ods/CMakeLists.txt
@@ -10,8 +10,7 @@ add_gdal_driver(
           PLUGIN_CAPABLE
           NO_DEPS)
 gdal_standard_includes(ogr_ODS)
-target_include_directories(ogr_ODS PRIVATE $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}>
-                                           $<TARGET_PROPERTY:ogr_MEM,SOURCE_DIR>)
+target_include_directories(ogr_ODS PRIVATE $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}>)
 add_executable(testparser EXCLUDE_FROM_ALL testparser.cpp)
 
 if (GDAL_USE_EXPAT)
