+++ ogr/ogrsf_frmts/openfilegdb/ogropenfilegdbdatasource.cpp
@@ -31,7 +31,7 @@
 #include "ogr_core.h"
 #include "ogr_feature.h"
 #include "ogr_geometry.h"
-#include "ogr_mem.h"
+#include "memdataset.h"
 #include "ogrsf_frmts.h"
 #include "ogr_swq.h"
